meetup: add flags for current data pulls
We want to be able to provide a cli flag the types of
data we want written to csv file. This Pr add the flag
`--analytics` which is a list of the datasets. it then parses
the list and queues up the corresponding functions.
