+++ meetup/meetup.go
@@ -4,12 +4,9 @@
 package meetup
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"log"
 	"net/http"
+	"strings"
 	"time"
 )
 
@@ -21,8 +18,14 @@ type Client struct {
 	url         string
 }
 
+var (
+	analyticsFuncs map[string]func()
+)
+
 func Setup(bearerToken string, proname string) Client {
-	return Client{
+	// TODO: add check for valid bearerToken
+	analyticsFuncs = make(map[string]func())
+	c := Client{
 		proname: proname,
 		ql: &http.Client{
 			Timeout: 10 * time.Second,
@@ -30,146 +33,22 @@ func Setup(bearerToken string, proname string) Client {
 		bearerToken: bearerToken,
 		url:         "https://api.meetup.com/gql",
 	}
+	analyticsFuncs["groups"] = c.GetGroupData
+	analyticsFuncs["eventRSVP"] = c.GetEventRSVPData
+	return c
 }
 
-type payloadql struct {
-	Query     string `json:"query"`
-	Variables string `json:"variables"`
-}
-
-type ProNetworkByUrlname struct {
-	Data struct {
-		ProNetwork struct {
-			GroupsSearch GroupsSearch `json:"groupsSearch,omitempty"`
-			EventsSearch EventsSearch `json:"eventsSearch,omitempty"`
-		} `json:"proNetworkByUrlname"`
-	} `json:"data"`
-}
-type GroupsSearch struct {
-	Count    int `json:"count"`
-	PageInfo struct {
-		HasNextPage bool   `json:"hasNextPage"`
-		StartCursor string `json:"startCursor"`
-		EndCursor   string `json:"endCursor"`
-	} `json:"pageInfo"`
-	Edges []Edge `json:"edges"`
-}
-
-type Edge struct {
-	Node Node `json:"node"`
-}
-type Node struct {
-	ID    string `json:"id"`
-	Title string `json:"title,omitempty"`
-	Name  string `json:"name,omitempty"`
-	Group struct {
-		ID   string `json:"id,omitempty"`
-		Name string `json:"name,omitempty"`
-	} `json:"group,omitempty"`
-	DateTime string `json:"dateTime,omitempty"`
-	Going    int    `json:"going,omitempty"`
-	Waiting  int    `json:"waiting,omitempty"`
-}
-
-type EventsSearch struct {
-	Count    int `json:"count"`
-	PageInfo struct {
-		HasNextPage bool   `json:"hasNextPage"`
-		StartCursor string `json:"startCursor"`
-		EndCursor   string `json:"endCursor"`
-	} `json:"pageInfo"`
-	Edges []Edge `json:"edges"`
-}
-
-// TODO(soypete): edit variables
-func getInputandVariables(isFirst bool, lastCursor, urlname string, numPerPage int) (string, string) {
-	if isFirst {
-		return "input: {first: $itemsNum}", fmt.Sprintf(`{"urlname":"%s","itemsNum": %d}`, urlname, numPerPage)
+func (m Client) GetAnalyticsFunc(funcs string) ([]func(), error) {
+	funcs = strings.Trim(funcs, "[]")
+	funcsList := strings.Split(funcs, ",")
+	var queuedFuncs []func()
+	for _, f := range funcsList {
+		key := strings.Trim(f, " ")
+		if _, ok := analyticsFuncs[key]; !ok {
+			return nil, fmt.Errorf("invalid flag: %s", key)
+			// TODO: return an error for invalid flad
+		}
+		queuedFuncs = append(queuedFuncs, analyticsFuncs[key])
 	}
-	return "input: {first: $itemsNum, after: $cursor}", fmt.Sprintf(`{"urlname":"%s","itemsNum": %d,"cursor": "%s"}`, urlname, numPerPage, lastCursor)
-
-}
-
-var queryTemplate = `query (%s) { proNetworkByUrlname(urlname: $urlname) { %s(%s) {count pageInfo { hasNextPage startCursor endCursor } edges { node { id %s } } } }}`
-
-func makePayloadql(isGroup, isfirst bool, lastCursor, urlname string, numPerPage int) payloadql {
-	variableTypes := "$urlname: String!, $itemsNum: Int!"
-	if !isfirst {
-		variableTypes = variableTypes + ", $cursor: String!"
-	}
-	searchType := `eventsSearch`
-	nodeQuery := `title group { id name } dateTime going waiting`
-	if isGroup {
-		searchType = `groupsSearch`
-		nodeQuery = `name`
-	}
-
-	input, variables := getInputandVariables(isfirst, lastCursor, urlname, 3)
-	query := fmt.Sprintf(queryTemplate, variableTypes, searchType, input, nodeQuery)
-	p := payloadql{
-		Query:     query,
-		Variables: variables,
-	}
-	return p
-}
-func (c Client) GetListOfGroups(cursor string) (ProNetworkByUrlname, error) {
-	isFirst := true
-	if cursor != "" {
-		isFirst = false
-	}
-	p := makePayloadql(true, isFirst, cursor, c.proname, 3)
-	body, err := c.sendRequest(p)
-	if err != nil {
-		return ProNetworkByUrlname{}, err
-	}
-	var respData ProNetworkByUrlname
-	err = json.Unmarshal(body, &respData)
-	if err != nil {
-		return ProNetworkByUrlname{}, err
-	}
-	return respData, nil
-}
-
-func (c Client) GetListOfEvents(cursor string) ProNetworkByUrlname {
-	isFirst := true
-	if cursor != "" {
-		isFirst = false
-	}
-	p := makePayloadql(false, isFirst, cursor, c.proname, 3)
-	body, err := c.sendRequest(p)
-	if err != nil {
-		log.Fatal(err)
-	}
-	// run it and capture the response
-	var respData ProNetworkByUrlname
-	err = json.Unmarshal(body, &respData)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return respData
-}
-
-func (c Client) sendRequest(ql payloadql) (resp []byte, err error) {
-	b, err := json.Marshal(ql)
-	if err != nil {
-		return nil, fmt.Errorf("cannot marshal graphql payload, %w", err)
-	}
-	req, err := http.NewRequest("POST", c.url, bytes.NewReader(b))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create new http.Request, %w", err)
-	}
-	// set header fields
-	req.Header.Add("Authorization", c.bearerToken)
-	req.Header.Add("Content-Type", "application/json")
-
-	res, err := c.ql.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed meetup.com api request, %w", err)
-	}
-	// TODO(soypete): check payload status
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read payload body, %w", err)
-	}
-	return body, nil
+	return queuedFuncs, nil
 }
